% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/time_calendar.R
\name{as.POSIXst}
\alias{POSIXst}
\alias{as.POSIXst}
\alias{c.POSIXst}
\title{Date-Time Objects in \pkg{stringi}}
\usage{
as.POSIXst(x, tz = attr(time, "tzone"), ...)

\method{c}{POSIXst}(..., recursive = FALSE)
}
\arguments{
\item{x}{...}

\item{tz}{\code{NULL} or \code{""} for the default time zone or
a single string with time zone identifier, see \code{\link{stri_timezone_list}}}

\item{recursive,...}{further arguments to be passed to or from other methods.}
}
\value{
t.b.d.
}
\description{
t.b.d.
}
\details{
An object of class \code{\link{POSIXst}},
inherits from (for compatibility with other base R functions)
\code{POSIXct} and \code{POSIX} classes.
In fact, it is a numeric vector representing the (signed) number of seconds
since the UNIX Epoch, i.e. 1970-01-01 00:00:00 UTC.
UTC leap seconds are ignored.

Thanks to this property, standard comparison operators, e.g. \code{<}, \code{==},
etc. or the \code{sort()} function may be used.

An object of class \code{\link{POSIXst}} may be equipped with
an attribute called \code{tzone}. Its value is used for date/time
formatting (e.g. when objects are printed in the console),
see \code{\link{format.POSIXst}} and \code{\link{stri_datetime_fields}}.
}
\seealso{
Other datetime: \code{\link{format.POSIXst}},
  \code{\link{stri_datetime_format}},
  \code{\link{stri_datetime_parse}};
  \code{\link{stri_datetime_add}},
  \code{\link{stri_datetime_add<-}};
  \code{\link{stri_datetime_create}};
  \code{\link{stri_datetime_fields}};
  \code{\link{stri_datetime_fstr}};
  \code{\link{stri_datetime_now}};
  \code{\link{stri_datetime_symbols}};
  \code{\link{stri_timezone_get}},
  \code{\link{stri_timezone_set}};
  \code{\link{stri_timezone_info}};
  \code{\link{stri_timezone_list}}
}

