% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/time_format.R
\name{stri_datetime_parse}
\alias{stri_datetime_parse}
\title{t.b.d.}
\usage{
stri_datetime_parse(str, format = "uuuu'-'MM'-'dd'T'HH':'mm':'ssxxx",
  tz = NULL, locale = NULL)
}
\arguments{
\item{str}{character vector}

\item{format}{single string, see \code{\link{stri_datetime_format}};
defaults to the ISO 8601 guideline}

\item{tz}{t.b.d}

\item{locale}{\code{NULL} or \code{""} for default locale,
or a single string with locale identifier; a non-Gregorian calendar
may be specified by setting \code{@calendar=name} keyword}
}
\value{
Returns an object of class \code{\link{POSIXst}}.
}
\description{
t,b,d,
}
\details{
t.b.d.
}
\examples{
stri_datetime_parse(c("2015-02-28", "2015-02-29"), "yyyy-MM-dd")
stri_datetime_parse("19 lipca 2015", "date_long", locale="pl_PL")
}
\seealso{
Other datetime: \code{\link{POSIXst}},
  \code{\link{as.POSIXst}}; \code{\link{format.POSIXst}},
  \code{\link{stri_datetime_format}};
  \code{\link{stri_datetime_add}},
  \code{\link{stri_datetime_add<-}};
  \code{\link{stri_datetime_create}};
  \code{\link{stri_datetime_fields}};
  \code{\link{stri_datetime_now}};
  \code{\link{stri_datetime_symbols}};
  \code{\link{stri_timezone_get}},
  \code{\link{stri_timezone_set}};
  \code{\link{stri_timezone_info}};
  \code{\link{stri_timezone_list}}
}

