% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding_conversion.R
\name{stri_enc_tonative}
\alias{stri_enc_tonative}
\title{Convert Strings To Native Encoding}
\usage{
stri_enc_tonative(str)
}
\arguments{
\item{str}{a character vector to be converted}
}
\value{
Returns a character vector.
}
\description{
Converts character strings with declared encodings
to the current native encoding.
}
\details{
This function just calls \code{\link{stri_encode}(str, NULL, NULL)}.
The current native encoding can be read with \code{\link{stri_enc_get}}.
Character strings declared to be in \code{bytes} encoding will fail here.

Note that if working in a UTF-8 environment,
resulting strings will be marked with \code{UTF-8}
and not \code{native}, see \code{\link{stri_enc_mark}}.
}
\seealso{
The official online manual of \pkg{stringi} at \url{https://stringi.gagolewski.com/}

Gagolewski M., \pkg{stringi}: Fast and portable character string processing in R, \emph{Journal of Statistical Software} 103(2), 2022, 1-59, \doi{10.18637/jss.v103.i02}

Other encoding_conversion:
\code{\link{about_encoding}},
\code{\link[=stri_enc_fromutf32]{stri_enc_fromutf32()}},
\code{\link[=stri_enc_toascii]{stri_enc_toascii()}},
\code{\link[=stri_enc_toutf32]{stri_enc_toutf32()}},
\code{\link[=stri_enc_toutf8]{stri_enc_toutf8()}},
\code{\link[=stri_encode]{stri_encode()}}
}
\concept{encoding_conversion}
\author{
\href{https://www.gagolewski.com/}{Marek Gagolewski} and other contributors
}
