% Generated by roxygen2 (4.0.2): do not edit by hand
\name{stri_in_fixed}
\alias{stri_in_fixed}
\title{Value Matching}
\usage{
stri_in_fixed(str, table, ...)
}
\arguments{
\item{str}{character vector of strings to search in}

\item{table}{character vector of values to be matched against}

\item{...}{further arguments passed to \code{\link{match}}}
}
\value{
Returns an integer vector.
}
\description{
For each element in \code{str}, this function returns
a position of the first matching element in \code{table}.
}
\details{
Vectorized over \code{str}.

This is a dummy wrapper for R's \code{\link{match}} function.
The only difference is that it ensures that the comparison
is made in UTF-8 (by calling \code{\link{stri_enc_toutf8}}).

It is (most likely) impossible to implement an algorithm
faster than the built-in \code{\link{match}}:
this function uses R's clever, hash table-based, string caching
system directly.
}
\seealso{
Other search_in: \code{\link{stringi-search}}
}

