% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{stri_isempty}
\alias{stri_isempty}
\title{Determine if a String is of Length Zero}
\usage{
stri_isempty(str)
}
\arguments{
\item{str}{character vector or an object coercible to}
}
\value{
Returns a logical vector of the same length as \code{str}.
}
\description{
This is the fastest way to find out
whether the elements of a character vector are empty strings.
}
\details{
Missing values are handled properly.
}
\examples{
stri_isempty(letters[1:3])
stri_isempty(c(',', '', 'abc', '123', '\u0105\u0104'))
stri_isempty(character(1))

}
\seealso{
The official online manual of \pkg{stringi} at \url{https://stringi.gagolewski.com/}

Gagolewski M., \pkg{stringi}: Fast and portable character string processing in R, \emph{Journal of Statistical Software} 103(2), 2022, 1-59, \doi{10.18637/jss.v103.i02}

Other length:
\code{\link{\%s$\%}},
\code{\link[=stri_length]{stri_length()}},
\code{\link[=stri_numbytes]{stri_numbytes()}},
\code{\link[=stri_pad_both]{stri_pad_both()}},
\code{\link[=stri_sprintf]{stri_sprintf()}},
\code{\link[=stri_width]{stri_width()}}
}
\concept{length}
\author{
\href{https://www.gagolewski.com/}{Marek Gagolewski} and other contributors
}
