% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stri_na2empty}
\alias{stri_na2empty}
\title{Replace NAs with Empty Strings}
\usage{
stri_na2empty(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
Returns a character vector.
}
\description{
This function replaces all missing values with empty strings.
See \code{\link{stri_replace_na}} for a generalization.
}
\examples{
stri_na2empty(c('a', NA, '', 'b'))

}
\seealso{
The official online manual of \pkg{stringi} at \url{https://stringi.gagolewski.com/}

Gagolewski M., \pkg{stringi}: Fast and portable character string processing in R, \emph{Journal of Statistical Software} 103(2), 2022, 1-59, \doi{10.18637/jss.v103.i02}

Other utils:
\code{\link[=stri_list2matrix]{stri_list2matrix()}},
\code{\link[=stri_remove_empty]{stri_remove_empty()}},
\code{\link[=stri_replace_na]{stri_replace_na()}}
}
\concept{utils}
\author{
\href{https://www.gagolewski.com/}{Marek Gagolewski} and other contributors
}
