% Generated by roxygen2 (4.0.0): do not edit by hand
\name{stri_num_rbnf_format}
\alias{stri_num_rbnf_format}
\title{t.b.d.}
\usage{
stri_num_rbnf_format(x, ruleset = "spellout", locale = NULL)
}
\arguments{
\item{x}{numeric vector}

\item{ruleset}{single string; one of \code{spellout}, \code{ordinal}, \code{duration},
\code{numbering_system}?}

\item{locale}{\code{NULL} or \code{""} for case mapping following
the conventions of the default locale, or a single string with
locale identifier, see \link{stringi-locale}.}
}
\value{
Returns a character vector
}
\description{
t.b.d.
}
\details{
t.b.d.
}
\seealso{
Other number_format: \code{\link{stri_num_rbnf_parse}}
}

