% Generated by roxygen2 (4.0.0): do not edit by hand
\name{stri_num_rbnf_parse}
\alias{stri_num_rbnf_parse}
\title{t.b.d.}
\usage{
stri_num_rbnf_parse(str, ruleset = "spellout", locale = NULL,
  lenient = TRUE)
}
\arguments{
\item{str}{character vector}

\item{ruleset}{single string; one of \code{spellout}, \code{ordinal}, \code{duration},
\code{numbering_system}?}

\item{locale}{\code{NULL} or \code{""} for case mapping following
the conventions of the default locale, or a single string with
locale identifier, see \link{stringi-locale}.}

\item{lenient}{single logical value; ....}
}
\value{
Returns a numeric vector (list of numeric vectors?)
}
\description{
t.b.d.
}
\details{
t.b.d.
}
\seealso{
Other number_format: \code{\link{stri_num_rbnf_format}}
}

