% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.R
\name{stri_order}
\alias{stri_order}
\title{Ordering Permutation}
\usage{
stri_order(str, decreasing = FALSE, na_last = TRUE, ..., opts_collator = NULL)
}
\arguments{
\item{str}{a character vector}

\item{decreasing}{a single logical value; should the sort order
be nondecreasing (\code{FALSE}, default)
or nonincreasing (\code{TRUE})?}

\item{na_last}{a single logical value; controls the treatment of \code{NA}s
in \code{str}. If \code{TRUE}, then missing values in \code{str} are put
at the end; if \code{FALSE}, they are put at the beginning;
if \code{NA}, then they are removed from the output}

\item{...}{additional settings for \code{opts_collator}}

\item{opts_collator}{a named list with \pkg{ICU} Collator's options,
see \code{\link{stri_opts_collator}}, \code{NULL}
for default collation options}
}
\value{
The function yields an integer vector that gives the sort order.
}
\description{
This function finds a permutation which rearranges the
strings in a given character vector into the ascending or descending
locale-dependent lexicographic order.
}
\details{
For more information on \pkg{ICU}'s Collator and how to tune it up
in \pkg{stringi}, refer to \code{\link{stri_opts_collator}}.

As usual in \pkg{stringi}, non-character inputs are coerced to strings,
see an example below for a somewhat non-intuitive behavior of lexicographic
sorting on numeric inputs.

This function uses a stable sort algorithm (\pkg{STL}'s \code{stable_sort}),
which performs up to \eqn{N*log^2(N)} element comparisons,
where \eqn{N} is the length of \code{str}.

For ordering with regards to multiple criteria (such as sorting
data frames by more than 1 column), see \code{\link{stri_rank}}.
}
\examples{
stri_order(c('hladny', 'chladny'), locale='pl_PL')
stri_order(c('hladny', 'chladny'), locale='sk_SK')

stri_order(c(1, 100, 2, 101, 11, 10))  # lexicographic order
stri_order(c(1, 100, 2, 101, 11, 10), numeric=TRUE)  # OK for integers
stri_order(c(0.25, 0.5, 1, -1, -2, -3), numeric=TRUE)  # incorrect
}
\references{
\emph{Collation} - ICU User Guide,
\url{https://unicode-org.github.io/icu/userguide/collation/}
}
\seealso{
The official online manual of \pkg{stringi} at \url{https://stringi.gagolewski.com/}

Gagolewski M., \pkg{stringi}: Fast and portable character string processing in R, \emph{Journal of Statistical Software} 103(2), 2022, 1-59, \doi{10.18637/jss.v103.i02}

Other locale_sensitive:
\code{\link{\%s<\%}},
\code{\link{about_locale}},
\code{\link{about_search_boundaries}},
\code{\link{about_search_coll}},
\code{\link[=stri_compare]{stri_compare()}},
\code{\link[=stri_count_boundaries]{stri_count_boundaries()}},
\code{\link[=stri_duplicated]{stri_duplicated()}},
\code{\link[=stri_enc_detect2]{stri_enc_detect2()}},
\code{\link[=stri_extract_all_boundaries]{stri_extract_all_boundaries()}},
\code{\link[=stri_locate_all_boundaries]{stri_locate_all_boundaries()}},
\code{\link[=stri_opts_collator]{stri_opts_collator()}},
\code{\link[=stri_rank]{stri_rank()}},
\code{\link[=stri_sort]{stri_sort()}},
\code{\link[=stri_sort_key]{stri_sort_key()}},
\code{\link[=stri_split_boundaries]{stri_split_boundaries()}},
\code{\link[=stri_trans_tolower]{stri_trans_tolower()}},
\code{\link[=stri_unique]{stri_unique()}},
\code{\link[=stri_wrap]{stri_wrap()}}
}
\concept{locale_sensitive}
\author{
\href{https://www.gagolewski.com/}{Marek Gagolewski} and other contributors
}
