% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stri_startswith_fixed}
\alias{stri_endswith_fixed}
\alias{stri_startswith_fixed}
\title{Determines if the Start or End of a String Matches a Fixed Pattern}
\usage{
stri_startswith_fixed(str, pattern)

stri_endswith_fixed(str, pattern)
}
\arguments{
\item{str}{character vector}

\item{pattern}{character vector}
}
\value{
Returns a logical vector.
}
\description{
These functions check if a string starts or ends with a fixed
pattern occurence.
}
\details{
Vectorized over \code{str} and \code{pattern}.

For natural language processing this function might not give
you desired results. Refer to \link{stringi-search-fixed} for more details.

If \code{pattern} is empty, then the result is \code{NA}
and a warning is generated.
}
\seealso{
Other search_fixed: \code{\link{stri_count_fixed}};
  \code{\link{stri_detect_fixed}};
  \code{\link{stri_locate_all_fixed}},
  \code{\link{stri_locate_first_fixed}},
  \code{\link{stri_locate_first_fixed}},
  \code{\link{stri_locate_last_fixed}},
  \code{\link{stri_locate_last_fixed}};
  \code{\link{stri_replace_all_fixed}},
  \code{\link{stri_replace_first_fixed}},
  \code{\link{stri_replace_first_fixed}},
  \code{\link{stri_replace_last_fixed}},
  \code{\link{stri_replace_last_fixed}};
  \code{\link{stri_split_fixed}};
  \code{\link{stringi-search-fixed}};
  \code{\link{stringi-search}}
}

