% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub.R
\name{stri_sub_list}
\alias{stri_sub_list}
\title{Extract or Replace Multiple Substrings}
\usage{
stri_sub_list(str, from = list(1L), to = list(-1L), length)
}
\arguments{
\item{str}{a character vector}

\item{from}{a list of integer vectors giving the start indexes or a
list of two-column matrices, each of type \code{cbind(from, to)}}

\item{to}{a list of integer vectors giving the end indexes}

\item{length}{a list of integer vectors giving the substring lengths}
}
\value{
Returns a list of character vectors.
}
\description{
\code{stri_sub_list} extracts multiple substrings from each string.

For extracting single substrings from each string, see
\code{\link{stri_sub}}.
}
\details{
Vectorized over \code{str}, \code{from} and
(\code{to} or \code{length}). Just like in \code{\link{stri_sub}}, parameters
\code{to} and \code{length} are mutually exclusive.

In one of the simplest scenarios, the i-th element of the resulting list
is generated by calling, e.g., \code{stri_sub(str[i], from[[i]], to[[i]])}.
As usual, if one of the inputs is of length smaller than the others,
recycling rule is applied.

If any of \code{from}, \code{to}, or \code{length} is not a list,
it is wrapped into a list.

\code{from} can be a list  of two-column matrices.
In such a case, the two columns play a role of \code{from} and \code{to},
respectively. Such types of index matrices are generated by
\code{\link{stri_locate_all}}.
If extraction based on \code{\link{stri_locate_first}}
or \code{\link{stri_locate_last}} is needed, see \code{\link{stri_sub}}.
}
\examples{
x <- c("12 3456 789", "abc", "", NA, "667")
stri_sub_list(x, stri_locate_all_regex(x, "[0-9]+")) # see stri_extract_all
stri_sub_list(x, stri_locate_all_regex(x, "[0-9]+", omit_no_match=TRUE))

}
\seealso{
Other indexing: \code{\link{stri_locate_all_boundaries}},
  \code{\link{stri_locate_all}}, \code{\link{stri_sub}}
}
\concept{indexing}
