% Generated by roxygen2 (4.0.2): do not edit by hand
\name{stri_subset_charclass}
\alias{stri_subset_charclass}
\title{Select Elements with a Character Class Match}
\usage{
stri_subset_charclass(str, pattern)
}
\arguments{
\item{str}{character vector to search in}

\item{pattern}{character vector specifying character classes to match,
see \link{stringi-search-charclass}}
}
\value{
Returns a character vector.
}
\description{
This function is roughly equivalent to a call to
\code{str[\link{stri_detect_charclass}(str, ...)]}.
}
\details{
Vectorized over \code{str} and \code{pattern}.

The resulting vector is of course always in UTF-8.
}
\examples{
\donttest{
stri_subset_charclass(c("stRRRingi","REXAMINE","123"),
   c("\\\\p{Ll}", "\\\\p{Lu}", "\\\\p{Zs}"))
}
}
\seealso{
Other search_charclass: \code{\link{stri_count_charclass}};
  \code{\link{stri_detect_charclass}};
  \code{\link{stri_extract_all_charclass}},
  \code{\link{stri_extract_first_charclass}},
  \code{\link{stri_extract_first_charclass}},
  \code{\link{stri_extract_last_charclass}},
  \code{\link{stri_extract_last_charclass}};
  \code{\link{stri_locate_all_charclass}},
  \code{\link{stri_locate_first_charclass}},
  \code{\link{stri_locate_first_charclass}},
  \code{\link{stri_locate_last_charclass}},
  \code{\link{stri_locate_last_charclass}};
  \code{\link{stri_replace_all_charclass}},
  \code{\link{stri_replace_first_charclass}},
  \code{\link{stri_replace_first_charclass}},
  \code{\link{stri_replace_last_charclass}},
  \code{\link{stri_replace_last_charclass}};
  \code{\link{stri_trim}}, \code{\link{stri_trim}},
  \code{\link{stri_trim_both}},
  \code{\link{stri_trim_left}},
  \code{\link{stri_trim_right}};
  \code{\link{stringi-search-charclass}};
  \code{\link{stringi-search}}

Other search_subset: \code{\link{stri_subset_coll}};
  \code{\link{stri_subset_fixed}};
  \code{\link{stri_subset_regex}};
  \code{\link{stri_subset}}; \code{\link{stringi-search}}
}

