% Generated by roxygen2 (4.0.2): do not edit by hand
\name{stri_subset_fixed}
\alias{stri_subset_fixed}
\title{Select Elements with a Fixed Pattern Match}
\usage{
stri_subset_fixed(str, pattern)
}
\arguments{
\item{str}{character vector}

\item{pattern}{character vector}
}
\value{
Returns a character vector.
}
\description{
This function is roughly equivalent to a call to
\code{str[\link{stri_detect_fixed}(str, ...)]}.
}
\details{
Vectorized over \code{str} and \code{pattern}.

For natural language processing this function might not give
you desired results. Refer to \link{stringi-search-fixed} for more details.

The resulting vector is of course always in UTF-8.
}
\examples{
\donttest{
stri_subset_fixed(c("stringi R", "REXAMINE", "123"), c('i', 'R', '0'))
stri_subset_fixed(c("stringi R", "REXAMINE", "123"), 'R')
}
}
\seealso{
Other search_fixed: \code{\link{stri_count_fixed}};
  \code{\link{stri_detect_fixed}};
  \code{\link{stri_endswith_fixed}},
  \code{\link{stri_startswith_fixed}};
  \code{\link{stri_locate_all_fixed}},
  \code{\link{stri_locate_first_fixed}},
  \code{\link{stri_locate_first_fixed}},
  \code{\link{stri_locate_last_fixed}},
  \code{\link{stri_locate_last_fixed}};
  \code{\link{stri_replace_all_fixed}},
  \code{\link{stri_replace_first_fixed}},
  \code{\link{stri_replace_first_fixed}},
  \code{\link{stri_replace_last_fixed}},
  \code{\link{stri_replace_last_fixed}};
  \code{\link{stri_split_fixed}};
  \code{\link{stringi-search-fixed}};
  \code{\link{stringi-search}}

Other search_subset: \code{\link{stri_subset_charclass}};
  \code{\link{stri_subset_coll}};
  \code{\link{stri_subset_regex}};
  \code{\link{stri_subset}}; \code{\link{stringi-search}}
}

