% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/time_calendar.R
\name{as.POSIXst}
\alias{POSIXst}
\alias{as.POSIXst}
\title{Date-Time Objects in \pkg{stringi}}
\usage{
as.POSIXst(x)
}
\arguments{
\item{x}{...}
}
\value{
t.b.d.
}
\description{
t.b.d.
}
\details{
An object of class \code{\link{POSIXst}},
inherits from (for compatibility with other base R functions)
\code{POSIXct} and \code{POSIX} classes.
In fact, it is a numeric vector representing the (signed) number of seconds
since the UNIX Epoch, i.e. 1970-01-01 00:00:00 UTC.
UTC leap seconds are ignored.

Thanks to this property, standard comparison operators, e.g. \code{<}, \code{==},
etc. or the \code{sort()} function may be used.
}
\seealso{
Other datetime: \code{\link{format.POSIXst}},
  \code{\link{stri_datetime_format}};
  \code{\link{stri_datetime_add}},
  \code{\link{stri_datetime_add<-}};
  \code{\link{stri_datetime_create}};
  \code{\link{stri_datetime_fields}};
  \code{\link{stri_datetime_now}};
  \code{\link{stri_datetime_parse}};
  \code{\link{stri_datetime_symbols}};
  \code{\link{stri_timezone_get}},
  \code{\link{stri_timezone_set}};
  \code{\link{stri_timezone_info}};
  \code{\link{stri_timezone_list}}
}

