% Generated by roxygen2 (4.0.0): do not edit by hand
\name{stri_count}
\alias{stri_count}
\title{Count the Number of Pattern Matches in a String}
\usage{
stri_count(str, ..., regex, fixed, coll, charclass)
}
\arguments{
\item{str}{character vector of strings to search in}

\item{...}{additional arguments passed to the underlying functions}

\item{regex}{character vector; regular expressions}

\item{fixed}{character vector; fixed patterns}

\item{coll}{character vector; canonically equivalent patterns}

\item{charclass}{character vector; identifiers of character classes}
}
\value{
Returns an integer vector.
}
\description{
A convenience function.
Calls either \code{\link{stri_count_regex}},
\code{\link{stri_count_fixed}}, \code{\link{stri_count_coll}},
or \code{\link{stri_count_charclass}},
depending on the argument used.
}
\details{
Unless you are very lazy, please call the underlying functions
directly for better performance.
}
\examples{
s <- "Lorem ipsum dolor sit amet, consectetur adipisicing elit."

\dontshow{if (stri_install_check(silent=TRUE))}
stri_count(s, fixed=letters)

\dontshow{if (stri_install_check(silent=TRUE))}
stri_count(s, regex="[[:alpha:]]")
}
\seealso{
Other search_count: \code{\link{stri_count_charclass}};
  \code{\link{stri_count_coll}};
  \code{\link{stri_count_fixed}};
  \code{\link{stri_count_regex}};
  \code{\link{stringi-search}}
}

