% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/time_calendar.R
\name{stri_datetime_add}
\alias{stri_datetime_add}
\title{Date and Time Arithmetic}
\usage{
stri_datetime_add(time, delta, units = "secs", calendar = "gregorian")
}
\arguments{
\item{time}{}

\item{delta}{}

\item{units}{}

\item{calendar}{}
}
\value{
Returns an object of class \code{StringiTime},
see \code{\link{stri_datetime_now}}.
}
\description{
Date and Time Arithmetic
}
\details{
Vectorized over \code{time} and \code{delta}

t.b.d.
}
\seealso{
Other datetime: \code{\link{stri_datetime_now}};
  \code{\link{stri_datetime_symbols}};
  \code{\link{stri_timezone_get}},
  \code{\link{stri_timezone_set}};
  \code{\link{stri_timezone_info}};
  \code{\link{stri_timezone_list}}
}

