% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/time_calendar.R
\name{stri_datetime_add}
\alias{stri_datetime_add}
\alias{stri_datetime_add<-}
\title{Date and Time Arithmetic}
\usage{
stri_datetime_add(time, value = 1L, units = "seconds", locale = NULL)

stri_datetime_add(time, units = "seconds", locale = NULL) <- value
}
\arguments{
\item{time}{a \code{POSIXct} object}

\item{value}{integer vector; signed number of units to add to a given time}

\item{units}{single string; one of \code{"years"}, \code{"months"},
\code{"weeks"}, \code{"days"}, \code{"hours"}, \code{"minutes"},
\code{"seconds"}, or \code{"milliseconds"}}

\item{locale}{\code{NULL} or \code{""} for default locale,
or a single string with locale identifier; a non-Gregorian calendar
may be specified by setting \code{@calendar=name} keyword}
}
\value{
Both functions return an object of class \code{\link{POSIXst}}.

The replacement version of \code{stri_datetime_add} modifies
the state of the \code{time} object.
}
\description{
Modifies a date-time object by adding a specific amount of time units.
}
\details{
Vectorized over \code{time} and \code{value}.

These date-time arithmetic operations are performed under the current
default time zone and locale. The \code{tzone} argument of \code{time}
is ignored. TO DO......................................................................


Note that e.g. January, 31 + 1 month = February, 28 or 29. TO DO ...........................
}
\examples{
x <- stri_datetime_now()
stri_datetime_add(x, units="months") <- 2
x
stri_datetime_add(x, -2, units="months")
stri_datetime_add(as.POSIXct("2014-04-20 12:00:00"), 1, units="years")
stri_datetime_add(as.POSIXct("2014-04-20 12:00:00"), 1, units="years", locale="@calendar=hebrew")
}
\references{
\emph{Calendar Classes} - ICU User Guide,
\url{http://userguide.icu-project.org/datetime/calendar}
}
\seealso{
Other datetime: \code{\link{POSIXst}},
  \code{\link{as.POSIXst}}; \code{\link{format.POSIXst}},
  \code{\link{stri_datetime_format}},
  \code{\link{stri_datetime_parse}};
  \code{\link{stri_datetime_create}};
  \code{\link{stri_datetime_fields}};
  \code{\link{stri_datetime_now}};
  \code{\link{stri_datetime_symbols}};
  \code{\link{stri_timezone_get}},
  \code{\link{stri_timezone_set}};
  \code{\link{stri_timezone_info}};
  \code{\link{stri_timezone_list}}
}

