% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/time_calendar.R
\name{stri_datetime_create}
\alias{stri_datetime_create}
\title{Crate a Date-Time Object}
\usage{
stri_datetime_create(year, month, day, hour = 12L, minute = 0L,
  second = 0, tz = NULL)
}
\arguments{
\item{year}{iteger vector; 0 is 1BC, -1 is 2BC, etc.}

\item{month}{iteger vector; months are 1-based}

\item{day}{iteger vector}

\item{hour}{iteger vector}

\item{minute}{iteger vector}

\item{second}{numeric vetor; fractional seconds are allowed}

\item{tz}{\code{NULL} or \code{""} for default time zone,
a single string with time zone ID otherwise.}
}
\value{
Returns an object of class \code{POSIXst},
which (for compatibility with other base R functions)
also inherits from \code{POSIXct} and \code{POSIX}.
In fact, it is a numeric vector representing the number of seconds
since the UNIX Epoch.
}
\description{
...
}
\details{
Vectorized over \code{year}, \code{month}, \code{hour},
\code{hour}, \code{minute}, and \code{second}.
}
\seealso{
Other datetime: \code{\link{stri_datetime_add}},
  \code{\link{stri_datetime_add<-}};
  \code{\link{stri_datetime_fields}};
  \code{\link{stri_datetime_now}};
  \code{\link{stri_datetime_symbols}};
  \code{\link{stri_timezone_get}},
  \code{\link{stri_timezone_set}};
  \code{\link{stri_timezone_info}};
  \code{\link{stri_timezone_list}}
}

