% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/time_calendar.R
\name{stri_datetime_create}
\alias{stri_datetime_create}
\title{Crate a Date-Time Object}
\usage{
stri_datetime_create(year, month, day, hour = 12L, minute = 0L,
  second = 0, tz = NULL, lenient = FALSE, locale = NULL)
}
\arguments{
\item{year}{iteger vector; 0 is 1BC, -1 is 2BC, etc.}

\item{month}{iteger vector; months are 1-based}

\item{day}{iteger vector}

\item{hour}{iteger vector}

\item{minute}{iteger vector}

\item{second}{numeric vetor; fractional seconds are allowed}

\item{tz}{\code{NULL} or \code{""} for the default time zone or
a single string with time zone ID otherwise}

\item{lenient}{single logical value; should the operation be lenient?}

\item{locale}{\code{NULL} or \code{""} for default locale,
or a single string with locale identifier; a non-Gregorian calendar
may be specified by setting \code{@calendar=name} keyword}
}
\value{
Returns an object of class \code{\link{POSIXst}}.
}
\description{
This function constructs date-time objects from numeric representations.
}
\details{
Vectorized over \code{year}, \code{month}, \code{hour},
\code{hour}, \code{minute}, and \code{second}.
}
\examples{
stri_datetime_create(2015, 12, 31, 23, 59, 59.999)
stri_datetime_create(5775, 8, 1, locale="@calendar=hebrew") # 1 Nisan 5775 -> 2015-03-21
stri_datetime_create(2015, 02, 29)
stri_datetime_create(2015, 02, 29, lenient=TRUE)
}
\seealso{
Other datetime: \code{\link{POSIXst}},
  \code{\link{as.POSIXst}}; \code{\link{format.POSIXst}},
  \code{\link{stri_datetime_format}},
  \code{\link{stri_datetime_parse}};
  \code{\link{stri_datetime_add}},
  \code{\link{stri_datetime_add<-}};
  \code{\link{stri_datetime_fields}};
  \code{\link{stri_datetime_fstr}};
  \code{\link{stri_datetime_now}};
  \code{\link{stri_datetime_symbols}};
  \code{\link{stri_timezone_get}},
  \code{\link{stri_timezone_set}};
  \code{\link{stri_timezone_info}};
  \code{\link{stri_timezone_list}}
}

