% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/time_calendar.R
\name{stri_datetime_fields}
\alias{stri_datetime_fields}
\title{Get Values for Date and Time Fields}
\usage{
stri_datetime_fields(time, locale = NULL)
}
\arguments{
\item{time}{a \code{POSIXct} object}

\item{locale}{\code{NULL} or \code{""} for default locale,
or a single string with locale identifier; a non-Gregorian calendar
may be specified by setting \code{@calendar=name} keyword}

\item{tz}{??????}
}
\value{
Returns a data frame with the following columns: TO DO...................................
\enumerate{
\item Years - 0 is 1BC, -1 is 2BC, etc.
\item Months ... months are 1-based
\item Days - ...
\item Hours - a single integer from 0 to 24,
\item Minutes - a single integer from 0 to 59
\item Seconds - a real number from 0 to 60 (fractional parts allowed)
\item TimeZone - ....
\item gmt offsett...
\item is dst...
\item week number...
\item day in year... ?
\item is weekend? ....
\item TO DO.....
}
}
\description{
Calculates and returns values for all date and time fields.
}
\details{
Vectorized over \code{time}.
}
\examples{
stri_datetime_fields(stri_datetime_now())
}
\seealso{
Other datetime: \code{\link{POSIXst}},
  \code{\link{as.POSIXst}}; \code{\link{format.POSIXst}},
  \code{\link{stri_datetime_format}},
  \code{\link{stri_datetime_parse}};
  \code{\link{stri_datetime_add}},
  \code{\link{stri_datetime_add<-}};
  \code{\link{stri_datetime_create}};
  \code{\link{stri_datetime_fstr}};
  \code{\link{stri_datetime_now}};
  \code{\link{stri_datetime_symbols}};
  \code{\link{stri_timezone_get}},
  \code{\link{stri_timezone_set}};
  \code{\link{stri_timezone_info}};
  \code{\link{stri_timezone_list}}
}

