% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/time_format.R
\name{stri_datetime_format}
\alias{format.POSIXst}
\alias{stri_datetime_format}
\title{Date and Time Formatting}
\usage{
stri_datetime_format(time, format = "uuuu-MM-dd'T'HH:mm:ssxxx", tz = NULL,
  locale = NULL)

\method{format}{POSIXst}(x, ..., usetz = TRUE)
}
\arguments{
\item{time}{a \code{POSIXct} object}

\item{format}{single string, see Details; defaults to a ISO 8601 guideline}

\item{tz}{t.b.d}

\item{locale}{\code{NULL} or \code{""} for default locale,
or a single string with locale identifier; a non-Gregorian calendar
may be specified by setting \code{@calendar=name} keyword}

\item{usetz}{this argument is purposely ignored}
}
\value{
Returns a character vector.
}
\description{
t.b.d
}
\details{
t.b.d
}
\examples{
# to do
}
\references{
\emph{Formatting Dates and Times} - ICU User Guide,
\url{http://userguide.icu-project.org/formatparse/datetime}
}
\seealso{
Other datetime: \code{\link{POSIXst}},
  \code{\link{as.POSIXst}};
  \code{\link{stri_datetime_add}},
  \code{\link{stri_datetime_add<-}};
  \code{\link{stri_datetime_create}};
  \code{\link{stri_datetime_fields}};
  \code{\link{stri_datetime_now}};
  \code{\link{stri_datetime_symbols}};
  \code{\link{stri_timezone_get}},
  \code{\link{stri_timezone_set}};
  \code{\link{stri_timezone_info}};
  \code{\link{stri_timezone_list}}
}

