% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/time_symbols.R
\name{stri_datetime_symbols}
\alias{stri_datetime_symbols}
\title{List Localizable Date-Time Formatting Data}
\usage{
stri_datetime_symbols(locale = NULL, context = "standalone",
  width = "wide")
}
\arguments{
\item{locale}{\code{NULL} or \code{""} for default locale,
or a single string with locale identifier.}

\item{context}{single string; one of: \code{"format"}, \code{"standalone"}.}

\item{width}{single string; one of: \code{"abbreviated"}, \code{"wide"}, \code{"narrow"}.}
}
\value{
Returns a list with the following named components:
\enumerate{
\item \code{Months} - month names,
\item \code{Weekdays} - weekday names,
\item \code{Quarters} - quarter names,
\item \code{AmPm} - AM/PM names,
\item \code{Era} - era names.
}
}
\description{
Returns a list of all localizable date-time formatting data,
including month and weeday names, localized AM/PM strings, etc.

\code{context} stands for a selector for date formatting context
and \code{width} - for  date formatting width.
}
\references{
\emph{DateFormatSymbols} class -- ICU API Documentation,
\url{http://icu-project.org/apiref/icu4c/classicu_1_1DateFormatSymbols.html}

\emph{Formatting Dates and Times} -- ICU User Guide,
\url{http://userguide.icu-project.org/formatparse/datetime}
}
\seealso{
Other datetime: \code{\link{stri_timezone_get}},
  \code{\link{stri_timezone_set}};
  \code{\link{stri_timezone_info}};
  \code{\link{stri_timezone_list}}
}

