% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{stri_dup}
\alias{stri_dup}
\title{Duplicate Strings}
\usage{
stri_dup(str, times)
}
\arguments{
\item{str}{a character vector of strings to be duplicated}

\item{times}{an integer vector with the numbers of times to duplicate each string}
}
\value{
Returns a character vector of the same length as \code{str}.
}
\description{
Duplicates each string \code{times} times
and concatenates the results.
}
\details{
Vectorized over \code{str} and \code{times}.
}
\examples{
stri_dup("a", 1:5)
stri_dup(c("a", NA, "ba"), 4)
stri_dup(c("abc", "pqrst"), c(4, 2))
}
\seealso{
Other join: 
\code{\link{\%s+\%}()},
\code{\link{stri_flatten}()},
\code{\link{stri_join_list}()},
\code{\link{stri_join}()}
}
\concept{join}
