% Generated by roxygen2 (4.0.2): do not edit by hand
\name{stri_enc_list}
\alias{stri_enc_list}
\title{List Known Character Encodings}
\usage{
stri_enc_list(simplify = FALSE)
}
\arguments{
\item{simplify}{single logical value; return a character vector or a
list of character vectors?}
}
\value{
If \code{simplify} is \code{FALSE} (the default), a list of
 character vectors is returned. Each list element represents a unique
 character encoding. The \code{name} attribute gives the \pkg{ICU} Canonical
 Name of an encoding family. The elements (character vectors) are
 its aliases.

If \code{simplify} is \code{TRUE}, then the resulting list
is coerced to a character vector and sorted, and returned with
removed duplicated entries.
}
\description{
Gives the list of encodings that are supported by \pkg{ICU}.
}
\details{
Please note that apart from given encoding identifiers and their aliases,
some other specifiers will be available in your conversion tasks.
This is due to the fact that \pkg{ICU} tries to normalize
converter names. E.g. \code{"UTF8"} is also valid,
see \link{stringi-encoding} for more information.
}
\seealso{
Other encoding_management: \code{\link{stri_enc_get}},
  \code{\link{stri_enc_set}}; \code{\link{stri_enc_info}};
  \code{\link{stri_enc_mark}};
  \code{\link{stringi-encoding}}
}

