% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stri_in_fixed}
\alias{stri_in_fixed}
\title{Value Matching}
\usage{
stri_in_fixed(str, table, nomatch = NA_integer_)
}
\arguments{
\item{str}{character vector of strings to search in}

\item{table}{character vector of values to be matched against}

\item{nomatch}{integer value to be returned in the case when no match is found}
}
\value{
Returns an integer vector.
}
\description{
This function returns a vector of the positions of (first) matches
of its first argument in its second.
}
\details{
Vectorized over \code{str}.
}
\seealso{
Other search_detect: \code{\link{stri_detect_charclass}};
  \code{\link{stri_detect_coll}};
  \code{\link{stri_detect_fixed}};
  \code{\link{stri_detect_regex}};
  \code{\link{stri_detect}}; \code{\link{stringi-search}}

Other search_regex: \code{\link{stri_count_regex}};
  \code{\link{stri_detect_regex}};
  \code{\link{stri_extract_all_regex}},
  \code{\link{stri_extract_first_regex}},
  \code{\link{stri_extract_first_regex}},
  \code{\link{stri_extract_last_regex}},
  \code{\link{stri_extract_last_regex}};
  \code{\link{stri_locate_all_regex}},
  \code{\link{stri_locate_first_regex}},
  \code{\link{stri_locate_first_regex}},
  \code{\link{stri_locate_last_regex}},
  \code{\link{stri_locate_last_regex}};
  \code{\link{stri_match_all_regex}},
  \code{\link{stri_match_first_regex}},
  \code{\link{stri_match_first_regex}},
  \code{\link{stri_match_last_regex}},
  \code{\link{stri_match_last_regex}};
  \code{\link{stri_opts_regex}};
  \code{\link{stri_replace_all_regex}},
  \code{\link{stri_replace_first_regex}},
  \code{\link{stri_replace_first_regex}},
  \code{\link{stri_replace_last_regex}},
  \code{\link{stri_replace_last_regex}};
  \code{\link{stri_split_regex}};
  \code{\link{stringi-search-regex}};
  \code{\link{stringi-search}}
}

