% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locale_management.R
\name{stri_locale_list}
\alias{stri_locale_list}
\title{List Available Locales}
\usage{
stri_locale_list()
}
\value{
Returns a character vector with locale identifiers
that are known to \pkg{ICU}.
}
\description{
Creates a character vector with all available locale identifies.
}
\details{
Note that some of the services may be unavailable in some locales.
Querying for locale-specific services is always performed
during the resource request.

See \link{stringi-locale} for more information.
}
\examples{
stri_locale_list()

}
\seealso{
The official online manual of \pkg{stringi} at \url{https://stringi.gagolewski.com/}

Gagolewski M., \pkg{stringi}: Fast and portable character string processing in R, \emph{Journal of Statistical Software} 103(2), 2022, 1-59, \doi{10.18637/jss.v103.i02}

Other locale_management: 
\code{\link{about_locale}},
\code{\link{stri_locale_info}()},
\code{\link{stri_locale_set}()}
}
\concept{locale_management}
\author{
\href{https://www.gagolewski.com/}{Marek Gagolewski} and other contributors
}
