% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locale_management.R
\name{stri_locale_list}
\alias{stri_locale_list}
\title{List Available Locales}
\usage{
stri_locale_list()
}
\value{
Returns a character vector with locale identifiers
that are known to \pkg{ICU}.
}
\description{
Creates a character vector with all known locale identifies.
}
\details{
Note that some of the services may be unavailable in some locales.
Querying for locale-specific services is always performed
during the resource request.

See \link{stringi-locale} for more information.
}
\seealso{
Other locale_management: 
\code{\link{stri_locale_info}()},
\code{\link{stri_locale_set}()},
\code{\link{stringi-locale}}
}
\concept{locale_management}
