% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_files.R
\name{stri_read_raw}
\alias{stri_read_raw}
\title{[DRAFT API] Read Whole Text File as Raw}
\usage{
stri_read_raw(fname)
}
\arguments{
\item{fname}{file name}
}
\value{
Returns a \code{raw}-type vector.
}
\description{
Reads a text file as-is, with no conversion or text line splitting.

\bold{[THIS IS AN EXPERIMENTAL FUNCTION]}
}
\details{
Once a text file is read into memory,
encoding detection (see \code{\link{stri_enc_detect2}}),
conversion (see \code{\link{stri_encode}}),
text lines split (see \code{\link{stri_split_lines1}})
can be performed.
}
\seealso{
Other files: 
\code{\link{stri_read_lines}()},
\code{\link{stri_write_lines}()}
}
\concept{files}
