% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stri_remove_empty}
\alias{stri_remove_empty}
\title{Remove all empty strings from a character vector}
\usage{
stri_remove_empty(x, na_empty = FALSE)
}
\arguments{
\item{x}{a character vector}

\item{na_empty}{should missing values be treated as empty strings?}
}
\value{
Always returns a character vector.
}
\description{
This function removes all empty strings from a character vector.
}
\examples{
stri_remove_empty(stri_na2empty(c("a", NA, "", "b")))
stri_remove_empty(c("a", NA, "", "b"))
stri_remove_empty(c("a", NA, "", "b"), TRUE)

}
\seealso{
Other utils: \code{\link{stri_list2matrix}},
  \code{\link{stri_na2empty}}
}
