% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_replace_4.R
\name{stri_replace_rstr}
\alias{stri_replace_rstr}
\title{Convert gsub-Style Replacement Strings}
\usage{
stri_replace_rstr(x)
}
\arguments{
\item{x}{character vector}
}
\value{
Returns a character vector.
}
\description{
Converts a \code{\link[base]{gsub}}-style replacement strings
to those which can be used in \code{\link{stri_replace}}.
In particular, \code{$} becomes \code{\\$} and \code{\\1} becomes \code{$1}.
}
\seealso{
The official online manual of \pkg{stringi} at \url{https://stringi.gagolewski.com/}

Other search_replace: 
\code{\link{about_search}},
\code{\link{stri_replace_all}()},
\code{\link{stri_trim_both}()}
}
\concept{search_replace}
\author{
\href{https://www.gagolewski.com/}{Marek Gagolewski} and other contributors
}
