% Generated by roxygen2 (4.0.2): do not edit by hand
\name{stri_startswith_fixed}
\alias{stri_endswith_fixed}
\alias{stri_startswith_fixed}
\title{Determine if the Start or End of a String Matches a Fixed Pattern}
\usage{
stri_startswith_fixed(str, pattern, from = 1L)

stri_endswith_fixed(str, pattern, to = -1L)
}
\arguments{
\item{str}{character vector}

\item{pattern}{character vector}

\item{from}{integer vector}

\item{to}{integer vector}
}
\value{
Returns a logical vector.
}
\description{
These functions check if a string starts or ends with a fixed
pattern occurence.
}
\details{
Vectorized over \code{str}, \code{pattern},
and \code{from} or \code{to}.

For natural language processing this function might not give
you desired results. Refer to \link{stringi-search-fixed} for more details.

If \code{pattern} is empty, then the result is \code{NA}
and a warning is generated.

Argument \code{start} controls the start position in \code{str}
at which \code{pattern} is mathed.
On the other hand, \code{to} gives the end position.

Indices given by \code{from} or \code{to} are 1-based,
i.e. an index equal to 1 denotes the first character
in a string, which gives a typical R look-and-feel.

For negative indices in \code{from} or \code{to}, counting starts
at the end of the string. E.g. index -1 denotes the last code point
in the string.
}
\seealso{
Other search_fixed: \code{\link{stri_locate_all_fixed}},
  \code{\link{stri_locate_first_fixed}},
  \code{\link{stri_locate_first_fixed}},
  \code{\link{stri_locate_last_fixed}},
  \code{\link{stri_locate_last_fixed}};
  \code{\link{stri_replace_all_fixed}},
  \code{\link{stri_replace_first_fixed}},
  \code{\link{stri_replace_first_fixed}},
  \code{\link{stri_replace_last_fixed}},
  \code{\link{stri_replace_last_fixed}};
  \code{\link{stringi-search-fixed}};
  \code{\link{stringi-search}}
}

