% Generated by roxygen2 (4.0.2): do not edit by hand
\name{stri_subset}
\alias{stri_subset}
\title{Select Elements That Match a Given Pattern}
\usage{
stri_subset(str, ..., regex, fixed, coll, charclass)
}
\arguments{
\item{str}{character vector of strings to search in}

\item{...}{additional arguments passed to the underlying functions}

\item{regex}{character vector; regular expressions}

\item{fixed}{character vector; fixed patterns}

\item{coll}{character vector; canonically equivalent patterns}

\item{charclass}{character vector; identifiers of character classes}
}
\value{
Returns a character vector.
}
\description{
A convenience function.
Calls either \code{\link{stri_subset_regex}},
\code{\link{stri_subset_fixed}}, \code{\link{stri_subset_coll}},
or \code{\link{stri_subset_charclass}},
depending on the argument used.
}
\details{
Unless you are very lazy, please call the underlying functions
directly for better performance.
}
\seealso{
Other search_subset: \code{\link{stri_subset_charclass}};
  \code{\link{stri_subset_coll}};
  \code{\link{stri_subset_fixed}};
  \code{\link{stri_subset_regex}};
  \code{\link{stringi-search}}
}

