% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stri_subset_fixed}
\alias{stri_subset_fixed}
\title{Subset Character Vector Using a Fixed Pattern Match}
\usage{
stri_subset_fixed(str, pattern)
}
\arguments{
\item{str}{character vector}

\item{pattern}{character vector}
}
\value{
Returns a character vector.
}
\description{
This function checks if there is at least one occurrence of a fixed pattern
in a string and if so it returns this vector element.
}
\details{
Vectorized over \code{str} and \code{pattern}.

For natural language processing this function might not give
you desired results. Refer to \link{stringi-search-fixed} for more details.

If \code{pattern} is empty, then the result is \code{NA}
and a warning is generated.
}
\examples{
\dontshow{if (stri_install_check(silent=TRUE))}
stri_subset_fixed(c("stringi R", "REXAMINE", "123"), c('i', 'R', '0'))

\dontshow{if (stri_install_check(silent=TRUE))}
stri_subset_fixed(c("stringi R", "REXAMINE", "123"), 'R')
}
\seealso{
Other search_fixed: \code{\link{stri_count_fixed}};
  \code{\link{stri_detect_fixed}};
  \code{\link{stri_endswith_fixed}},
  \code{\link{stri_startswith_fixed}};
  \code{\link{stri_locate_all_fixed}},
  \code{\link{stri_locate_first_fixed}},
  \code{\link{stri_locate_first_fixed}},
  \code{\link{stri_locate_last_fixed}},
  \code{\link{stri_locate_last_fixed}};
  \code{\link{stri_replace_all_fixed}},
  \code{\link{stri_replace_first_fixed}},
  \code{\link{stri_replace_first_fixed}},
  \code{\link{stri_replace_last_fixed}},
  \code{\link{stri_replace_last_fixed}};
  \code{\link{stri_split_fixed}};
  \code{\link{stringi-search-fixed}};
  \code{\link{stringi-search}}

Other search_subset: \code{\link{stri_subset_regex}};
  \code{\link{stri_subset}}
}

