% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stri_subset_regex}
\alias{stri_subset_regex}
\title{Subset Character Vector Using a Regex Pattern Match}
\usage{
stri_subset_regex(str, pattern)
}
\arguments{
\item{str}{character vector of strings to search in}

\item{pattern}{character vector of regular expressions}

\item{opts_regex}{a named list with \pkg{ICU} Regex settings
as generated with \code{\link{stri_opts_regex}}; \code{NULL}
for default settings}
}
\value{
Returns a character vector.
}
\description{
This function checks if there is at least one occurrence of a regex pattern
in a string and if so it returns this vector element.
}
\details{
Unless you are very lazy, please call the underlying functions
directly for better performance.

Vectorized over \code{str} and \code{pattern}.

If \code{pattern} is empty, then the result is \code{NA}
and a warning is generated.
}
\examples{
\dontshow{if (stri_install_check(silent=TRUE))}
stri_subset_regex(c("stringi R", "REXAMINE", "123"), c('i', 'R', '0'))

\dontshow{if (stri_install_check(silent=TRUE))}
stri_subset_regex(c("stringi R", "REXAMINE", "123"), '[a-z]')
}
\seealso{
Other search_detect: \code{\link{stri_detect_charclass}};
  \code{\link{stri_detect_coll}};
  \code{\link{stri_detect_fixed}};
  \code{\link{stri_detect_regex}};
  \code{\link{stri_detect}}; \code{\link{stri_in_fixed}};
  \code{\link{stri_subset_fixed}};
  \code{\link{stringi-search}}

Other search_regex: \code{\link{stri_count_regex}};
  \code{\link{stri_detect_regex}};
  \code{\link{stri_extract_all_regex}},
  \code{\link{stri_extract_first_regex}},
  \code{\link{stri_extract_first_regex}},
  \code{\link{stri_extract_last_regex}},
  \code{\link{stri_extract_last_regex}};
  \code{\link{stri_in_fixed}};
  \code{\link{stri_locate_all_regex}},
  \code{\link{stri_locate_first_regex}},
  \code{\link{stri_locate_first_regex}},
  \code{\link{stri_locate_last_regex}},
  \code{\link{stri_locate_last_regex}};
  \code{\link{stri_match_all_regex}},
  \code{\link{stri_match_first_regex}},
  \code{\link{stri_match_first_regex}},
  \code{\link{stri_match_last_regex}},
  \code{\link{stri_match_last_regex}};
  \code{\link{stri_opts_regex}};
  \code{\link{stri_replace_all_regex}},
  \code{\link{stri_replace_first_regex}},
  \code{\link{stri_replace_first_regex}},
  \code{\link{stri_replace_last_regex}},
  \code{\link{stri_replace_last_regex}};
  \code{\link{stri_split_regex}};
  \code{\link{stringi-search-regex}};
  \code{\link{stringi-search}}
}

