% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/time_zone.R
\name{stri_timezone_info}
\alias{stri_timezone_info}
\title{Query Given Time Zone}
\usage{
stri_timezone_info(tz = NULL, locale = NULL)
}
\arguments{
\item{tz}{\code{NULL} or \code{""} for default time zone,
a single string with time zone ID otherwise.}

\item{locale}{\code{NULL} or \code{""} for default locale,
or a single string with locale identifier.}
}
\value{
Returns a list with the following named components:
\enumerate{
\item \code{ID} (time zone identifier),
\item \code{Name} (localized human-readable time zone name),
\item \code{RawOffset} (raw GMT offset, in hours, before taking daylight savings int account), and
\item \code{UsesDaylightTime} (states whether a time zone uses daylight savings time
 in the current Gregorian calendar year).
}
}
\description{
Provides some basic information on a given time zone identifier.
}
\details{
With this function you may obtain some basic information
on any supported time zone.

For more information on time zone representation in \pkg{ICU},
see \code{\link{stri_timezone_list}}.
}
\examples{
stri_timezone_info()
stri_timezone_info(locale="sk_SK")
}
\seealso{
Other datetime: \code{\link{POSIXst}},
  \code{\link{as.POSIXst}};
  \code{\link{stri_datetime_add}},
  \code{\link{stri_datetime_add<-}};
  \code{\link{stri_datetime_create}};
  \code{\link{stri_datetime_fields}};
  \code{\link{stri_datetime_now}};
  \code{\link{stri_datetime_symbols}};
  \code{\link{stri_timezone_get}},
  \code{\link{stri_timezone_set}};
  \code{\link{stri_timezone_list}}

Other timezone: \code{\link{stri_timezone_get}},
  \code{\link{stri_timezone_set}};
  \code{\link{stri_timezone_list}}
}

