% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/time_zone.R
\name{stri_timezone_list}
\alias{stri_timezone_list}
\title{List Available Time Zone Identifiers}
\usage{
stri_timezone_list(region = NA_character_, offset = NA_integer_)
}
\arguments{
\item{region}{single string;
a ISO 3166 two-letter country code or UN M.49 three-digit area code}

\item{offset}{single numeric value;
a given raw offset from GMT, in hours}
}
\value{
Returns a character vector.
}
\description{
Returns a list of available time zone identifiers.

A time zone represents an offset applied to Greenwich Mean Time (GMT)
to obtain local time (GMT is similar, but not precisely identical,
to Universal Coordinated Time, or UTC. In \pkg{ICU} the two terms
are used interchangeably since \pkg{ICU} does not concern itself with
either leap seconds or historical behavior).
The offset might vary throughout the year,
if daylight savings time (DST) is used, or might be the same all year long.
Typically, regions closer to the equator do not use DST.
If DST is in use, then specific rules define the point at which
 the offset changes and the amount by which it changes.

If  \code{offset} and \code{region} are \code{NA} (the default),
all time zones are returned. Otherwise,
only time zone IDs with a given raw offset from GMT are given
and/or time zones corresponding to a given region are provided.
Note that the effect of daylight savings time is ignored.
}
\examples{
stri_timezone_list()
stri_timezone_list(offset=1)
stri_timezone_list(offset=5.5)
stri_timezone_list(offset=5.75)
stri_timezone_list(region="PL")
}
\references{
\emph{TimeZone} class -- ICU API Documentation,
\url{http://www.icu-project.org/apiref/icu4c/classicu_1_1TimeZone.html}

\emph{ICU 4.8 Time Zone Names}.
\url{http://site.icu-project.org/design/formatting/timezone/icu-4-8-time-zone-names}

\emph{ICU TimeZone classes} -- ICU User Guide,
\url{http://userguide.icu-project.org/datetime/timezone}
}
\seealso{
Other datetime: \code{\link{stri_datetime_add}},
  \code{\link{stri_datetime_add<-}};
  \code{\link{stri_datetime_now}};
  \code{\link{stri_datetime_symbols}};
  \code{\link{stri_timezone_get}},
  \code{\link{stri_timezone_set}};
  \code{\link{stri_timezone_info}}

Other timezone: \code{\link{stri_timezone_get}},
  \code{\link{stri_timezone_set}};
  \code{\link{stri_timezone_info}}
}

