% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/time_zone.R
\name{stri_timezone_get}
\alias{stri_timezone_get}
\alias{stri_timezone_set}
\title{Set or Get Default Time Zone in \pkg{stringi}}
\usage{
stri_timezone_get()

stri_timezone_set(tz)
}
\arguments{
\item{tz}{single string; time zone identifier}
}
\value{
\code{stri_timezone_set} returns a string with
previously used timezone, invisibly.

\code{stri_timezone_get} returns a single string
with the current default time zone.
}
\description{
\code{stri_timezone_set} changes default time zone for all  functions
in the \pkg{stringi} package,
i.e. establishes the meaning of the ``\code{NULL} time zone'' argument
of date/time processing functions. On the other hand, \code{stri_timezone_get}
gets current default time zone.

For more information on time zone representation in \pkg{ICU},
see \code{\link{stri_timezone_list}}.
}
\details{
Unless the default time zone has already been set using \code{stri_timezone_set},
the default time zone is determined by querying the system using methods
in ICU's \code{TPlatformUtilities}. If the system routines fail,
or if they specify a TimeZone or TimeZone offset which is not recognized,
the TimeZone indicated by the ID kLastResortID is instantiated and made the default.
}
\examples{
\dontrun{
oldtz <- stri_timezone_set("Europe/Warsaw")
# ... some time zone-dependent operations
stri_timezone_set(oldtz) # restore previous default time zone
}
}
\references{
\emph{TimeZone} class -- ICU API Documentation,
\url{http://www.icu-project.org/apiref/icu4c/classicu_1_1TimeZone.html}
}
\seealso{
Other datetime: \code{\link{POSIXst}},
  \code{\link{as.POSIXst}};
  \code{\link{stri_datetime_add}},
  \code{\link{stri_datetime_add<-}};
  \code{\link{stri_datetime_create}};
  \code{\link{stri_datetime_fields}};
  \code{\link{stri_datetime_now}};
  \code{\link{stri_datetime_symbols}};
  \code{\link{stri_timezone_info}};
  \code{\link{stri_timezone_list}}

Other timezone: \code{\link{stri_timezone_info}};
  \code{\link{stri_timezone_list}}
}

