% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_other.R
\name{stri_trans_char}
\alias{stri_trans_char}
\title{Translate Characters}
\usage{
stri_trans_char(str, pattern, replacement)
}
\arguments{
\item{str}{character vector}

\item{pattern}{a single character string providing code points to be translated}

\item{replacement}{a single character string giving translated code points}
}
\value{
Returns a character vector.
}
\description{
Translates Unicode code points in each input string.
}
\examples{
stri_trans_char("id.123", ".", "_")
stri_trans_char("babaab", "ab", "01")
stri_trans_char("accctggaaactg", "actg", "tgac")
}
\seealso{
Other transform: \code{\link{stri_trans_general}},
  \code{\link{stri_trans_list}},
  \code{\link{stri_trans_nfc}},
  \code{\link{stri_trans_tolower}}
}
\concept{transform}
