% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trans_other.R
\name{stri_trans_char}
\alias{stri_trans_char}
\title{Translate characters}
\usage{
stri_trans_char(str, pattern, replacement)
}
\arguments{
\item{str}{character vector}

\item{pattern}{character vector providing characters to be translated}

\item{replacement}{character vector providing translated characters}
}
\value{
Returns a character vector.
}
\description{
Translates character in each input string.
}
\details{
Vectorized over \code{str}, \code{pattern} and \code{replacement}.

If \code{pattern} and its corresponding \code{replacement} are of different
lenghts, then the extra characters in the longer of the two are ignored.

If characters in some \code{pattern} are not unique, last corresponding
replacement is used.
}
\examples{
stri_trans_char("id.123", ".", "_")
stri_trans_char("babaab", "ab", "01")
}
\seealso{
Other transform: \code{\link{stri_trans_general}};
  \code{\link{stri_trans_isnfc}},
  \code{\link{stri_trans_isnfd}},
  \code{\link{stri_trans_isnfkc}},
  \code{\link{stri_trans_isnfkc_casefold}},
  \code{\link{stri_trans_isnfkd}},
  \code{\link{stri_trans_nfc}},
  \code{\link{stri_trans_nfd}},
  \code{\link{stri_trans_nfkc}},
  \code{\link{stri_trans_nfkc_casefold}},
  \code{\link{stri_trans_nfkd}};
  \code{\link{stri_trans_list}};
  \code{\link{stri_trans_tolower}},
  \code{\link{stri_trans_totitle}},
  \code{\link{stri_trans_toupper}}
}

