% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trans_other.R
\name{stri_trans_char}
\alias{stri_trans_char}
\title{Translate Characters}
\usage{
stri_trans_char(str, pattern, replacement)
}
\arguments{
\item{str}{character vector}

\item{pattern}{a single character string providing code points to be translated}

\item{replacement}{a single character string giving translated code points}
}
\value{
Returns a character vector.
}
\description{
Translates Unicode code points in each input string.
}
\details{
Vectorized over \code{str} and with respect to each code point
in \code{pattern} and \code{replacement}.

If \code{pattern} and \code{replacement} consist of a different number
of code points, then the extra code points in the longer of the two are ignored,
with a warning.

If code points in a given \code{pattern} are not unique,
last corresponding replacement code point is used.
}
\examples{
stri_trans_char("id.123", ".", "_")
stri_trans_char("babaab", "ab", "01")
}
\seealso{
Other transform: \code{\link{stri_trans_general}};
  \code{\link{stri_trans_isnfc}},
  \code{\link{stri_trans_isnfd}},
  \code{\link{stri_trans_isnfkc}},
  \code{\link{stri_trans_isnfkc_casefold}},
  \code{\link{stri_trans_isnfkd}},
  \code{\link{stri_trans_nfc}},
  \code{\link{stri_trans_nfd}},
  \code{\link{stri_trans_nfkc}},
  \code{\link{stri_trans_nfkc_casefold}},
  \code{\link{stri_trans_nfkd}};
  \code{\link{stri_trans_list}};
  \code{\link{stri_trans_tolower}},
  \code{\link{stri_trans_totitle}},
  \code{\link{stri_trans_toupper}}
}

