% Generated by roxygen2 (4.0.2): do not edit by hand
\name{stri_trans_list}
\alias{stri_trans_list}
\title{List Available Text Transforms and Transliterators}
\usage{
stri_trans_list()
}
\value{
Returns a character vector.
}
\description{
Returns a list of available text transform identifiers.
Each of them may be used in \code{\link{stri_trans_general}}
tasks.
}
\references{
\emph{General Transforms} -- ICU User Guide,
\url{http://userguide.icu-project.org/transforms/general}
}
\seealso{
Other transform: \code{\link{stri_trans_general}};
  \code{\link{stri_trans_isnfc}},
  \code{\link{stri_trans_isnfd}},
  \code{\link{stri_trans_isnfkc}},
  \code{\link{stri_trans_isnfkc_casefold}},
  \code{\link{stri_trans_isnfkd}},
  \code{\link{stri_trans_nfc}},
  \code{\link{stri_trans_nfd}},
  \code{\link{stri_trans_nfkc}},
  \code{\link{stri_trans_nfkc_casefold}},
  \code{\link{stri_trans_nfkd}};
  \code{\link{stri_trans_tolower}},
  \code{\link{stri_trans_totitle}},
  \code{\link{stri_trans_toupper}}
}

